% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhtmlEchoLifecycle.R
\name{draw}
\alias{draw}
\title{rhtmlEchoLifecycle HTML Widget}
\source{
https://github.com/Displayr/rhtmlEchoLifecycle
}
\usage{
draw(width = 600, height = 600, settings = "")
}
\description{
A HTMLWidget that echos function calls for debugging purposes.
}
\examples{

rhtmlEchoLifecycle::draw(600, 600, '')

}
\author{
Kyle Zeeuwen <kyle.zeeuwen@gmail.com>
}
